skip (remove this commit) - temporary fix build
To enable us to look at the preview, with a placeholder page. I have
updated the "KubeCon!" navbar so it links to PROJECT-MEMBERS.md, as an
external source. When we merge KUBECON.md in fluxcd/community#349 we can
revert this commit or rebase it out.

We may edit the structure further before all that happens, in previous
years we had multiple tabs on the Google Sites page and we might want to
replicate that as the "Project🔽" dropdown style - for now I have only
made one page, because there aren't as many Flux talks as previous years
that I'm aware of.

Signed-off-by: Kingdon P Barrett <kingdon@tuesdaystudios.com>